发现一个bug，假如没有上传文件的话，我也会执行删除文件操作，现在加了一个判断，没有上传文件的话，就不去后台删除文件

+++ web/src/views/AcList.vue
@@ -83,7 +83,10 @@ export default {
       )
         .then(async () => {
           await this.$http.delete(`/rest/Ac/${row._id}`);
-          await this.$http.delete(`/deleteFile/${row.fileName}`);
+          if (row.fileName == " " || row.fileName == undefined) {
+            //如果根本上传文件，就不用去后台删除文件了。
+            await this.$http.delete(`/deleteFile/${row.fileName}`);
+          }
           this.fetch();
           this.$notify({
             title: "成功",
